requirements.txt
gradio
pandas
requests
sqlite3
python-dotenv
